PR#10 Modification
created Activity+Ext, to include the extensions for activities, and add here the transitionFragment function
