+++ app/src/main/java/com/example/bankodemia/UI/View/MainActivity.kt
@@ -2,7 +2,10 @@ package com.example.bankodemia.UI.View
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.fragment.app.Fragment
+import com.example.bankodemia.Core.transitionFragment
 import com.example.bankodemia.R
+import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
